修复编译错误，项目成功启动
- story-logic pom: volcengine SDK版本改为1.0.16（本地缓存），fastjson2补全版本号
- ArenaService: power字段类型long→修正Math.max类型兼容
- RebirthService: lambda捕获变量加final，修复effectively-final编译报错

+++ logic/arena-logic/src/main/java/com/iohao/mmo/arena/service/ArenaService.java
@@ -36,8 +36,8 @@ public class ArenaService {
         ArenaPlayer opponent = getArenaInfo(opponentId);
 
         // 基于战力计算胜率：胜率 = playerPower / (playerPower + opponentPower)
-        int playerPower = Math.max(player.getPower(), 1);
-        int opponentPower = Math.max(opponent.getPower(), 1);
+        long playerPower = Math.max(player.getPower(), 1L);
+        long opponentPower = Math.max(opponent.getPower(), 1L);
         double winChance = (double) playerPower / (playerPower + opponentPower);
         boolean win = Math.random() < winChance;
